requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.12
asttokens==2.0.4
attrs==19.3.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.4
blinker==1.4
Brotli==1.0.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clyent==1.2.2
colorama==0.4.3
colorlover==0.3.0
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
cryptography==2.9.2
cufflinks==0.17.3
cycler==0.10.0
dash~=1.16.1
dash-auth==1.3.2
dash-bootstrap-components==0.10.3
dash-core-components==1.10.1
dash-enterprise-auth==0.0.4
dash-html-components==1.0.3
dash-renderer==1.5.0
dash-table==4.8.1
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
email-validator==1.1.1
entrypoints==0.3
executing==0.5.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.9.0
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SeaSurf==0.2.2
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
glob2==0.7
gunicorn==20.0.4
icecream==2.0.0
idna==2.9
importlib-metadata==1.6.0
ipdb==0.13.3
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.4
kiwisolver==1.2.0
libarchive-c==2.9
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.2.2
menuinst==1.4.16
mistune==0.8.4
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.21
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.1
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.5
olefile==0.46
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
plotly~=4.8.2
PriceIndices==1.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
psycopg2==2.8.5
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.10.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
redis==3.5.3
requests==2.23.0
retrying==1.3.3
ruamel-yaml==0.15.87
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
sshtunnel==0.1.5
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
ua-parser==0.10.0
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
WTForms==2.3.1
xlrd==1.2.0
xmltodict==0.12.0
zipp==3.1.0

python-dotenv~=0.14.0
pip~=20.2.3
zlib~=1.2.11
wheel~=0.34.2
openssl~=1.1.1g
selenium~=3.141.0
fire~=0.3.1
yaml~=0.1.7
termcolor~=1.1.0
django~=3.1.1
setuptools~=47.1.1
backports~=1.0
winpty~=0.4.3
asgiref~=3.2.10
jupyter_client~=6.1.3
prompt_toolkit~=3.0.5
pycares~=3.1.1
ipython_genutils~=0.2.0
jupyter_core~=4.6.3
prometheus_client~=0.7.1
extract~=1.0
pandoc~=2.2.3.2
python~=3.7.6
uritemplate~=3.0.1
google-api-python-client~=1.12.1
importlib_metadata~=1.6.0
pyorbital~=1.6.0
pymongo~=3.11.0